fix: locate `compiler.jar` using Node's regular mechanism.
Newer npm versions change the installation location of node modules. Dependent
modules are not always installed into subfolders, in particular if there are
no version conflicts. That reduces file system usage, but also means this code
needs to be a bit more standard in how it locates the `compiler.jar`.